feat(status): adding item-status component
